[Starts #162780453] Move the models and views to their specific directories
